First test wave: All OK except: "/api/exemplares/transferir"   : NOK sp_TransferirExemplar não existe MapPost "/api/nucleos/"   : NOK Refazer swagger para colocar token de autenticação "/api/nucleos/transferencia"  : NOK Refazer swagger para colocar token de autenticação "/api/nucleos/dashboard"  : NOK Refazer swagger para colocar token de autenticação

+++ XPTOBusiness/Models/Nucleos.cs
@@ -8,7 +8,7 @@ namespace XPTOBusiness.Models
 {
     public class Nucleo
     {
-        public int ID_Nucleo { get; set; }
+        public long ID_Nucleo { get; set; }
         public string? Nome { get; set; }
         public string? Local { get; set; }
         public byte ID_TipoNucleo { get; set; }
